+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Require VSCode 1.18.
 * Add support for catkin tools alongside catkin_make (thanks to @JamesGiller).
 * Remove some unimplemented commands.
+* Add "ROS: Create Terminal" command.
 
 ## 0.1.0